Refactored the code that deals with non-square pixels.
This involved changes to the API, and to the behavior of the command-line
utility.

The code that calculates the new density used to be awkwardly split between
the library and the command-line util. It's now entirely in the command-line
util.

Added the -noresize command line option, to prevent IW from fixing up
non-square pixels.

Fixed at least one bug: images with nonsquare pixels could end up the wrong
shape in some cases, for example when using relative sizing.